+++ apps/io-services-cms-webapp/src/config.ts
@@ -126,6 +126,11 @@ export const QueueConfig = t.type({
 });
 export type QueueConfig = t.TypeOf<typeof QueueConfig>;
 
+// Application Insight configuration
+export const ApplicationInsightConfig = t.type({
+  APPINSIGHTS_INSTRUMENTATIONKEY: NonEmptyString,
+});
+
 // Services pagination configuration
 export const PaginationConfig = t.type({
   PAGINATION_DEFAULT_LIMIT: withDefault(
@@ -147,6 +152,14 @@ const ServiceIdQualityCheckExclusionList = t.type({
   ),
 });
 
+const FeatureFlags = t.type({
+  // UserId List allowed to sync services from CMS to Legacy
+  USERID_CMS_TO_LEGACY_SYNC_INCLUSION_LIST: withDefault(
+    CommaSeparatedListOf(NonEmptyString),
+    []
+  ),
+});
+
 // Global app configuration
 export type IConfig = t.TypeOf<typeof IConfig>;
 export const IConfig = t.intersection([
@@ -164,7 +177,13 @@ export const IConfig = t.intersection([
     QueueConfig,
     ServiceIdQualityCheckExclusionList,
   ]),
-  t.intersection([CosmosLegacyConfig, PaginationConfig, JiraLegacyProjectName]),
+  t.intersection([
+    CosmosLegacyConfig,
+    PaginationConfig,
+    JiraLegacyProjectName,
+    ApplicationInsightConfig,
+    FeatureFlags,
+  ]),
 ]);
 
 export const envConfig = {