+++ apps/io-services-cms-webapp/src/main.ts
@@ -14,7 +14,6 @@ import {
   SubscriptionCIDRsModel,
 } from "@pagopa/io-functions-commons/dist/src/models/subscription_cidrs";
 import * as O from "fp-ts/Option";
-import * as RTE from "fp-ts/ReaderTaskEither";
 import * as RA from "fp-ts/ReadonlyArray";
 import * as RR from "fp-ts/ReadonlyRecord";
 import { pipe } from "fp-ts/lib/function";
@@ -46,6 +45,8 @@ import { handler as onServiceHistoryHandler } from "./watchers/on-service-histor
 import { handler as onServiceLifecycleChangeHandler } from "./watchers/on-service-lifecycle-change";
 import { handler as onServicePublicationChangeHandler } from "./watchers/on-service-publication-change";
 import { createWebServer } from "./webservice";
+
+import { initTelemetryClient } from "./utils/applicationinsight";
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires, @typescript-eslint/no-unused-vars
 const BASE_PATH = require("../host.json").extensions.http.routePrefix;
@@ -83,6 +84,11 @@ const fsmPublicationClient = ServicePublication.getFsmClient(
   servicePublicationStore
 );
 
+// AppInsights client for Telemetry
+const telemetryClient = initTelemetryClient(
+  config.APPINSIGHTS_INSTRUMENTATIONKEY
+);
+
 const legacyServicesContainer = cosmosdbClient
   .database(config.LEGACY_COSMOSDB_NAME)
   .container(SERVICE_COLLECTION_NAME);
@@ -98,6 +104,7 @@ export const httpEntryPoint = pipe(
     fsmLifecycleClient,
     fsmPublicationClient,
     subscriptionCIDRsModel,
+    telemetryClient,
   },
   createWebServer,
   expressToAzureFunction
@@ -191,8 +198,7 @@ export const onLegacyServiceChangeEntryPoint = pipe(
 );
 
 export const onServiceHistoryChangeEntryPoint = pipe(
-  onServiceHistoryHandler,
-  RTE.fromReaderEither,
+  onServiceHistoryHandler(config, apimClient),
   processBatchOf(ServiceHistory),
   setBindings((results) => ({
     requestSyncLegacy: pipe(