+++ apps/io-services-cms-webapp/src/webservice/controllers/__test__/delete-service.test.ts
@@ -10,6 +10,7 @@ import {
   SubscriptionCIDRsModel,
 } from "@pagopa/io-functions-commons/dist/src/models/subscription_cidrs";
 import { UserGroup } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/azure_api_auth";
+import { setAppContext } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/context_middleware";
 import { NonNegativeInteger } from "@pagopa/ts-commons/lib/numbers";
 import {
   IPatternStringTag,
@@ -80,6 +81,19 @@ const containerMock = {
 } as unknown as Container;
 
 const subscriptionCIDRsModel = new SubscriptionCIDRsModel(containerMock);
+
+const mockAppinsights = {
+  trackEvent: vi.fn(),
+  trackError: vi.fn(),
+} as any;
+
+const mockContext = {
+  log: {
+    error: vi.fn((_) => console.error(_)),
+    info: vi.fn((_) => console.info(_)),
+  },
+} as any;
+
 describe("deleteService", () => {
   afterEach(() => {
     vi.restoreAllMocks();
@@ -92,7 +106,10 @@ describe("deleteService", () => {
     fsmLifecycleClient,
     fsmPublicationClient,
     subscriptionCIDRsModel,
+    telemetryClient: mockAppinsights,
   });
+
+  setAppContext(app, mockContext);
 
   const aService = {
     id: "aServiceId",
@@ -124,6 +141,7 @@ describe("deleteService", () => {
       .set("x-user-id", anUserId)
       .set("x-subscription-id", aManageSubscriptionId);
 
+    expect(mockContext.log.error).toHaveBeenCalledOnce();
     expect(response.statusCode).toBe(404);
   });
 
@@ -141,6 +159,7 @@ describe("deleteService", () => {
       .set("x-user-id", anUserId)
       .set("x-subscription-id", aManageSubscriptionId);
 
+    expect(mockContext.log.error).toHaveBeenCalledOnce();
     expect(response.statusCode).toBe(409);
   });
 
@@ -170,6 +189,7 @@ describe("deleteService", () => {
       .set("x-user-id", anUserId)
       .set("x-subscription-id", aManageSubscriptionId);
 
+    expect(mockContext.log.error).not.toHaveBeenCalled();
     expect(response.statusCode).toBe(204);
   });
 
@@ -185,6 +205,7 @@ describe("deleteService", () => {
       .set("x-user-id", aDifferentUserId)
       .set("x-subscription-id", aDifferentManageSubscriptionId);
 
+    expect(mockContext.log.error).toHaveBeenCalledOnce();
     expect(response.statusCode).toBe(403);
   });