+++ apps/io-services-cms-webapp/src/webservice/controllers/review-service.ts
@@ -1,4 +1,5 @@
 import { ApiManagementClient } from "@azure/arm-apimanagement";
+import { Context } from "@azure/functions";
 import { ServiceLifecycle } from "@io-services-cms/models";
 import { SubscriptionCIDRsModel } from "@pagopa/io-functions-commons/dist/src/models/subscription_cidrs";
 import {
@@ -14,6 +15,7 @@ import {
   ClientIp,
   ClientIpMiddleware,
 } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/client_ip_middleware";
+import { ContextMiddleware } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/context_middleware";
 import { RequiredBodyPayloadMiddleware } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/required_body_payload";
 import { RequiredParamMiddleware } from "@pagopa/io-functions-commons/dist/src/utils/middlewares/required_param";
 import {
@@ -24,38 +26,33 @@ import {
   checkSourceIpForHandler,
   clientIPAndCidrTuple as ipTuple,
 } from "@pagopa/io-functions-commons/dist/src/utils/source_ip_check";
+import { initAppInsights } from "@pagopa/ts-commons/lib/appinsights";
 import {
-  IResponseErrorConflict,
-  IResponseErrorForbiddenNotAuthorized,
-  IResponseErrorInternal,
-  IResponseErrorNotFound,
-  IResponseErrorTooManyRequests,
   IResponseSuccessNoContent,
   ResponseSuccessNoContent,
 } from "@pagopa/ts-commons/lib/responses";
 import { NonEmptyString } from "@pagopa/ts-commons/lib/strings";
-import { pipe } from "fp-ts/lib/function";
 import * as TE from "fp-ts/lib/TaskEither";
+import { pipe } from "fp-ts/lib/function";
 import { IConfig } from "../../config";
 import { ReviewRequest as ReviewRequestPayload } from "../../generated/api/ReviewRequest";
 import { fsmToApiError } from "../../utils/converters/fsm-error-converters";
+import { ErrorResponseTypes, getLogger } from "../../utils/logger";
 import { serviceOwnerCheckManageTask } from "../../utils/subscription";
+
+const logPrefix = "ReviewServiceHandler";
 
 type Dependencies = {
   config: IConfig;
   fsmLifecycleClient: ServiceLifecycle.FsmClient;
   apimClient: ApiManagementClient;
+  telemetryClient: ReturnType<typeof initAppInsights>;
 };
 
-type HandlerResponseTypes =
-  | IResponseSuccessNoContent
-  | IResponseErrorForbiddenNotAuthorized
-  | IResponseErrorNotFound
-  | IResponseErrorConflict
-  | IResponseErrorTooManyRequests
-  | IResponseErrorInternal;
+type HandlerResponseTypes = IResponseSuccessNoContent | ErrorResponseTypes;
 
 type ReviewServiceHandler = (
+  context: Context,
   auth: IAzureApiAuthorization,
   clientIp: ClientIp,
   attrs: IAzureUserAttributesManage,
@@ -68,15 +65,16 @@ export const makeReviewServiceHandler =
     config,
     fsmLifecycleClient: fsmLifecycleClient,
     apimClient,
+    telemetryClient,
   }: Dependencies): ReviewServiceHandler =>
-  (_auth, __, attrs, serviceId, body) =>
+  (context, auth, __, ___, serviceId, body) =>
     pipe(
       serviceOwnerCheckManageTask(
         config,
         apimClient,
         serviceId,
-        _auth.subscriptionId,
-        _auth.userId
+        auth.subscriptionId,
+        auth.userId
       ),
       TE.chainW((sId) =>
         pipe(
@@ -87,6 +85,24 @@ export const makeReviewServiceHandler =
           TE.mapLeft(fsmToApiError)
         )
       ),
+      TE.map((resp) => {
+        telemetryClient.trackEvent({
+          name: "api.manage.services.review",
+          properties: {
+            userSubscriptionId: auth.subscriptionId,
+            serviceId,
+            autoPublish: body.auto_publish,
+          },
+        });
+        return resp;
+      }),
+      TE.mapLeft((err) =>
+        getLogger(context, logPrefix).logErrorResponse(err, {
+          userSubscriptionId: auth.subscriptionId,
+          serviceId,
+          autoPublish: body.auto_publish,
+        })
+      ),
       TE.toUnion
     )();
 
@@ -94,8 +110,11 @@ export const applyRequestMiddelwares =
   (subscriptionCIDRsModel: SubscriptionCIDRsModel) =>
   (handler: ReviewServiceHandler) => {
     const middlewaresWrap = withRequestMiddlewares(
+      // extract the Azure functions context
+      ContextMiddleware(),
       // only allow requests by users belonging to certain groups
       AzureApiAuthMiddleware(new Set([UserGroup.ApiServiceWrite])),
+      // extract the client IP from the request
       ClientIpMiddleware,
       // check manage key
       AzureUserAttributesManageMiddleware(subscriptionCIDRsModel),
@@ -107,7 +126,9 @@ export const applyRequestMiddelwares =
     return wrapRequestHandler(
       middlewaresWrap(
         // eslint-disable-next-line max-params
-        checkSourceIpForHandler(handler, (_, c, u, __, ___) => ipTuple(c, u))
+        checkSourceIpForHandler(handler, (_, __, c, u, ___, ____) =>
+          ipTuple(c, u)
+        )
       )
     );
   };